requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
pillow
tk
streamlit 